Move reset class types to a job
Fixes ZPS-959

Halt installation if zenjobs is running

update error message.  no need to log message
